refactor: Move mode switcher to shared components, remove analytics overview and expense categories pages, and delete unused nav collapsible and API type.
