+++ src/components/shared/mode-switcher.tsx
