+++ src/components/sidebar/data.ts
@@ -21,10 +21,8 @@ export const navigationTitle = {
 
 	// Analytics pages
 	Analytics: "Analytics",
-	AnalyticsOverview: "Overview",
 	AnalyticsAvailableCars: "Available Cars",
 	AnalyticsSoldCars: "Sold Cars",
-	AnalyticsExpenseCategories: "Expense Categories",
 	AnalyticsCarProfitSummary: "Car Profit Summary",
 	AnalyticsSharersTotal: "Sharers Total",
 
@@ -68,20 +66,12 @@ export const data: {
 			icon: ChartColumn,
 			items: [
 				{
-					title: navigationTitle.AnalyticsOverview,
-					url: paths.dashboard.analytics.overview.getHref(),
-				},
-				{
 					title: navigationTitle.AnalyticsAvailableCars,
 					url: paths.dashboard.analytics.availableCars.getHref(),
 				},
 				{
 					title: navigationTitle.AnalyticsSoldCars,
 					url: paths.dashboard.analytics.soldCars.getHref(),
-				},
-				{
-					title: navigationTitle.AnalyticsExpenseCategories,
-					url: paths.dashboard.analytics.expenseCategories.getHref(),
 				},
 				{
 					title: navigationTitle.AnalyticsCarProfitSummary,
@@ -121,10 +111,6 @@ export const titleMap = [
 		title: navigationTitle.AnalyticsSoldCars,
 	},
 	{
-		match: paths.dashboard.analytics.expenseCategories.getHref(),
-		title: navigationTitle.AnalyticsExpenseCategories,
-	},
-	{
 		match: paths.dashboard.analytics.carProfitSummary.getHref(),
 		title: navigationTitle.AnalyticsCarProfitSummary,
 	},
@@ -136,10 +122,6 @@ export const titleMap = [
 	// ------------------------
 	// MAIN PAGES (lower priority)
 	// ------------------------
-	{
-		match: paths.dashboard.analytics.overview.getHref(),
-		title: navigationTitle.AnalyticsOverview,
-	},
 	{
 		match: paths.dashboard.cars.getHref(),
 		title: navigationTitle.CarsListing,
