--- src/types/api.ts
@@ -1,5 +1,0 @@
-export type Meta = {
-	page: number;
-	total: number;
-	totalPages: number;
-};
